Tweak gscms interface to avoid use of globals.
Previously the only way to implement gscms_get_clrtname was to use a
static buffer and to return a pointer to it. We change the interface
so that gscms_get_clrtname takes a memory pointer, and allocates
space for the returned thing.

The LCMS2 implementation of this still uses a 256 byte stack based
buffer as part of it's operation (so could potentially have problems
with colorants longer than 256 bytes), but there is no way to avoid
this in the current LCMS interface.