Fix windows compilation of pcl3 devices.
The pcl3 devices include support for a pagecount module that doesn't
work under windows. The documentation for these devices suggests that
they can be made to work by predefining EPRN_NO_PAGECOUNTFILE and
editing the dependencies in the makefile to remove the need for
pagefile.c.

This commit simplifies that by making that symbol be defined
automatically on a windows build, and nobbling pagefile.c with a