Add in support for spot color overprinting with standard CMYK devices
Currently only the psdcmyk and tiffsep devices provide simulation of spot
color overprinting.   With the following commit it is possible to get
simulation of overprinting with other CMYK based devices.  This is done
using the command line option -dSimulateOverprint.   The simulation is
achieved by having the overprint device blend the CMYK colorant with what
ever has currently been drawn.