Fix -Zb decode flags.
Some of the debugging code that had been changed to provide memory
pointers to the debugging print functions was attempting to
retrive a memory pointer from a stream that was on the stack.
Stack based streams have null memory pointers, resulting in SEGVs
when the pointers were used.

Revert these calls back to the non threadsafe ones.