Enable default spot name usage when output profile is NCLR
Previously I had added the requirement that we used something like
-sICCOutputColors="Cyan, Magenta, Yellow, Black, Orange, Violet"
when using an NCLR ICC profile for the device profile (valid only for
tiffsep and psdcmyk devices).  This was to ensure proper naming of the
profile colorants.   It was pointed out by a customer that they would
prefer to have a set of default names so this was now added.
If -sICCOutputColors is not specified and we are using an NCLR ICC profile,
then we will have spot colors from the profile named ICC_COLOR_0, ICC_COLOR_1
etc.