ps2write - while stripping OtherSubrs, don't try to index off stack
The code assmued that a Callsubr operator woudl be preceded by the subr index
but in this case the subr index is provided by another Subr (crazy font...).

All we need to do is cdheck the stack index, and if its 0, not index from the
stack but simply write the Callsubr operator.

Fixed Bug #693325

Expected Differences
tpc2.ps should no longer seg fault on the cluster with ps2write.