Fix interpolation code to behave consistently in the presence of pdf14.
There are special cases in the setup of the image interpolation code
designed to kick in in the presence of halftoning devices and either
limit the amount or change the nature of the interpolation done.

The test for this is confused by the presence of a pdf14 device though.

We move the special cases down into a device specific op call to avoid
this. This also means that devices can have finer control over exactly
when these operations are used.

One possibility would be to expose them as device params, but I haven't
done that here yet.