Support PJL language string HP-GL/2.
The language string HP-GL/2 now selects the "RTL" mode in PCL.  It
should also be used to select the language HP-GL/2 upon language
switch, but only one language string is supported per emulation, and
the feature is not really needed.  Any HPGL/2 rtl file should select
the PCL emulation anyway because of the character of the input stream
and the HP-GL/2 PJL command will select the RTL personality.