pdfwrite/ps2write - enable 64-bit file handling
Subsequent to commit c81c8f1b8740c484ec332080b5ce5f718357694d it is now
possible to have 64-bit offsets returned by stell, and this makeds it possible,
for the first time, to write PDF files > 4GB.

There's a fair amount of reworking in here, but we can now produce very large
temporary files, and from that produce very large PDF and PostScript output.

There are some restrictions; no stream can exceed 2GB and no font can exceed 2GB

As might be expected this is *not* well tested as we don't actually have any
test files which exercise this. I have tested PDF and PostScript file production
with a single file which produces a 10GB temporary file, a 7 GB PDF file and a 6GB
PostScript file.

The resulting PDF file opens with Adobe Acrobat and can Ghostscript, and the
PostScript file can be consumed by Adobe Distiller and Ghostscript.

Nevertheless, bugs are possible.

No differences expected as this is not lcuster tested.