Fix tiffsep and introduce tiffscaled4 device.
The tiffsep device was supposed to be able to be called with
-dBitsPerComponent set to 1,2,4 or 8 to allow differing bit
depths of output. This setting would work for the separation
planes, but would not produce a composite when bpc was not 8
due to the code not supporting that.

We add code here to generate composites in 1,2 and 4 bit modes
too, along with some sanity checking of the compression mode.

We also add a tiffscaled4 mode where we produce 1bpc cmyk from
a downscaled dither of 8bpc cmyk.