Bug 693166: Speed images through the clist.
In investigating bug 693166 Ray spotted that the calculation of which
bands were touched by images was slack in the presence of a clipping
path - he proposed a simple patch to fix this.

Unfortunately it had a knock on effect where vertical offsets could be
introduced into the topmost band. This is fixed here by a second small
change in image_band_box.

Cluster testing shows 10 small changes, all well within the usual
clist differences.