Bug 693430: missing "fallback" for DecodeParams entry
In a ResuableStreamDecode filter with an array of subfilters (/Filter entry
in the parameter dictionary). If teh DecodeParams array also exists, and has
one or more entries set to "null", we weren't pushing an empty dictionary
onto the stack for the subsequent call to the filter operator.

The equivalent code for a totally absent DecodeParams array was already
correct.

No cluster differences.