Hin-Tak Leung's fix for bug 693413 - large output file size due to images being decimated to rectangles.
The fix allows any orthogonal image, given the color information can
be properly represented in a PXL, to be converted directly to a PXL
image.  The change has not been cluster tested, the PXL device is
tested in the overnight testing and there is no practical means to
test the device's code change with all test files locally, so the
commit also serves as a test submission.