Fix part of bug #693415.  HPGL's command to create a sub-polygon closed paths unconditionally.
Thanks to Norbert Janssen for providing a fix to conditionally close
the path only if the current pen is down.  The fix revealed another
problem where PCL was not creating a new sub-path when the current pen
was up and drawing a circle (CI).  This is remedied with issuing a
"moveto" not "gapto" when starting the circle.  Currently the only
means of doing a moveto explicitly is to temporarily disable the "hpgl
path mode" while calling the gs_moveto() procedure.  This arrangement
will likely be improved in the future by adding a gapto function to
the graphics library language interface.

These changes result in fixing the original source problem and several
improvements in the PCL CET.  A few new minor discrepancies
(regressions) were found in the CET's and another test (all documented
in the bug), these were either inconsequential or deferrable and we
feel this change is substantial progress and should not be held up for
the other issues to be addressed.