Fix for issues with icc output intent
Proper rendering of the output rendering intent Ghent test files to RGB based devices require that
the output intent profile is used as a proofing profile.  Unfortunately littleCMS does not allow one
to specify different rendering intents to use when connecting a group of profiles together.  This
is needed to render these files correctly.   The solution was to create a device link profile for the
first rendering intent (e.g perceptual) when going from source to proof.  Then the remainder of the
transforms are handled in a relative colorimetric manner from proof to destination and possibly through
the device's link profile.