Bug #693111 fix refinement.
Refine the fix 8bec29de to add a configuration option so individual
commands can be selected for the unusual behavior of being ignored in
raster mode.  I don't think we have a solid understanding of this
problem and this makes the fix significantly less likely to result in
a regression.