pdfwrite - further TrueType composite glyph improvement. Bug #693444
When we encounter a TrueType composite glyph, we add the 2 component glyphs
to the table of names, so that we generate a correct CMAP subtable in the
embedded subset font.

However, the name is based on the GID of the glyph, and if the GID happens
match a name in the font Encoding (type 42 font) *and* we use the glyph
which really corresponds to that name (but obviously has a different GID)
then we would end up with the wrong CMAP generated.

By not removing the GS_MIN_GLYPH_INDEX bias we create a new name which is
numeric and based on the GID. This guarantees no name collisions.

This leads to a regression in an XPS test file, but I think it is of the
'wrong before, differently wrong now' category and propose to ignore it on
that basis. (its a change in spacing, which cannot be directly affected by
this change.