pdfwrite - properly set the PreserveSMask parameter for non-transparent output
The PDF interpreter has historically handled SMasks by passing them directly to
the device, when the device is pdfwrite. This was altered to use a device
parameter rather than checking the3 device name.

This did not alter the functionality, which has always bee incorrect when
pdfwrite is not preserving transparency. We now alter the value of the
device parameter if we are instead rendering the transparent content to an
image. SMasks should now behave correctly when using PDF/A-1, PDF/X or
when setting the CompatibilityLevel to less than 1.4.

No differences expected as this is not cluster tested.