Enable proper blending to occur when we draw transparency patterns.
This fixes  Bug 693498.  Previously we were not getting the blending information to
the group that we push to draw into.   Also,  there was a difference in how the clist
and non-clist pattern cases had to be handled.  In the clist case, we do not know
what the blend mode for a group encompassing the pattern would be until the pattern is
run through the clist.   Hence its logic with respect to transparency required a change.
Also fixed a bug that occurred when we used the -Zv debug option.