Bug 693490: Search all XObject resources.
Adobe Acrobat appears to search all nested XObjects for resource names
contrary to the PDF specification. This patch follows the reference
implementation and saves a pointer to XObject resource dictionary on
the graphic state. When a resource cannot be found in a standard way,
the stack of graphic states is searched.