pdfwritre - change fixed buffer for CIDSystemInfo to dynamic allocation
Bug #693514 "Broken PDF produced from PostScript by Windows 8 Microsoft
PS Class Driver"

The problem is that the CIDSystemInfo Registry and Ordering produced by
the Microsoft XPS to PS device is > 32 characters. The code to write the
CIDSystemInfo used a fixed buffer of 32 characters. If that overflowed we
returned a limitcheck which caused pdfwrite to attempt a type 3 fallback
which also failed.

Changed the code to (more logically) allocate a buffer for the temporary
usage. Also if this fails we return a VMerror which doesn't trigger an
attempt to fall back to type 3.

No Expected Differences