Bug 693538: tweak TTF cmap format 4 parsing.
Empty segments in the cmap table (that is, those where the start character code
is the same as the end code) aren't ignored, we correctly "fake" a one entry
segment. It seems, however, that we should *not* do that when the start and
end codes are zero - in that case, we should ignore it.

No cluster differences.