Fix "fallback" to know about Microtype fonts
In the event that FAPI fails to handle a font, there is a (temporary) fallback
to the AFS code.

That fallback *must* not be used for the Microtype fonts loaded via UFST, since
the font loading code uses the error return to determine that we've loaded all
the fonts from the FCO.

No cluster differences.