Fix bug 693361: Rendering threads unsafe use of main thread allocator.
The PCL main thread allocator is not thread safe, and the rendering
threads would end up using this allocator to form link profiles (in
the lcms logic) because the device profile had the 'memory' allocator
set to the main thread's allocator.

The fix found by Robin Watts was to use the thread's device (which
have a copy of the device profile with the thread's memory allocator,
set by gs_putdeviceparams) in the call to create_buf_device.