Bug 693541: Fix planar depth calculation from color_info.
At various places in the code, we assume that for a planar device:

 dev->color_info.depth = dev->color_info.num_components * plane_depth

This is NOT necessarily the case (due to tag planes etc).

We therefore avoid the need for this calculation by adjusting the
return value from gxdso_is_native_planar; rather than returning 0
or 1 (for not-planar or planar), we now return 0 or the depth of
the planes.

The logic in pdf14_create_clist_device is updated to use this. This
gets the file mentioned in the bug through to another (seemingly
unrelated) crash that I will investigate next.