Fix bug 693557: SEGV regression with NumRenderingThreads > 0 in pattern_load
This was due to the pdf14 device changing num_components during clist playback
when it had not changed during writing the clist. When a pattern tried to load
it failed the check that the depth of the pattern matched the depth of the
device. This was traced to picking the pdf14cmyk device during rendering instead
of the pdf14CMYKspot device because the thread's max_components was set to 4,
not 14 (as during clist writing). Fix the MaxSeparations in gsdparam to return
max_components instead of num_components, then fix devn_get_params to pass the
PageSpotColors param and the setting of num_components in devn_put_params to
ignore max_separations since that doesn't make sense.