Bug 693541: Fix color setup in pdf14_create_clist_device.
When we create a pdf14 clist device (a pdf14cmykspot device, in this
case) in pdf14_create_clist_device, we are at pains to set the
num_components and max_components appropriately. At no point do we
similarly sanitise the depth however.

The fix applied here is an exact parallel to the first commit for
this bug.