CUPS Raster "cups" device: Improvements for pages size matching with PPD
These changes improve the matching of the sizes of the incoming pages with
the page sizes available for thr printer according to the PPD file.

Instead of stopping to searcg through the PPDs page sizes when a size
gets found which fits within the limits, we always go through all page
sizes to find the best match. We consider the best match the size with
the name given on the command line/by the input file and matching the
size of the current page within the limits, second level is a match of
size and margins and third level is only matching the size. If there
is more than one match of the same level, the product of the
horizontal and vertival size deviation counts.

In addition, we let the gstoraster wrapper CUPS filter for Ghostscript
call Ghostscript with all parameters on the command line (especially
also the page size name) also when the input is PostScript.