Properly initialise 'clip' device
Bug #693575 "segfault with ps2write and -dDisablePAPI"

Although labeled as a PS writer problem this is in fact a general bug. In
gx_image_fill_masked we call gx_image_fill_masked_start() with a pointer
to a gx_device * which we expect to be filled in.

This will be the device which accumulates the clip, if the current target
device can accumulate clips then we use it, otherwise we set up an
accumulation device.

However there was a path through gx_image_fill_masked_start which did not
set the device, leading to the variable being left uninitialised and
causing a later crash.

This patch initialises the variable to the current device before calling
gx_image_fill_masked_start and also fixes the code path in that routine
so that it always fills in the device. Just to be safe.

No differences expected.