Update int_freed_top in the correct chunk
In i_free_object() we were identifying the correct chunk for the memory
address we were freeing, but then always updating int_freed_top in the
current chunk. This could cause us to fail to consolidate free blocks when
we possibly could, or to attempt to do so when it would be pointless.

So we now identify whether the address is in the current chunk, and write
int_freed_top in that, or use the actual chunk pointer for the freed
memory block.

No cluster differences.