Bug 693571: properly close allocator chunk during save
During a save operation, we "close" the open chunk and save the state of the
current generation of an allocator before creating a new generation and opening
a new chunk for the new generation.

After doing that, we try to eliminate any redundant changes listed for the
"saved" allocator instance. That can end freeing memory in the "closed" chunk,
leaving it in an unstable state, and causing the garbage collector to try to
validate already freed memory.

In drop_redundant_changes() we now call alloc_open_chunk() and
alloc_close_chunk() so that, in event we do disard changes (thus freeing
memory), the allocator is left in an internally consistent state, and
avoids confusing the chunk validation code during garbage collection.

No cluster differences.