Fix bug #693586 - clist error message.
The error was caused by sending image data to the library after the
page was output, now graphics mode is flushed and closed before the
page is output.  The bug was introduced by ea54a03 which closed
graphics mode at the time of "end job" which could occur after output
page.  We do need to leave that commit intact as the job could end
without outputting a page, and the raster state would not be cleaned
up.