Work around a make problem with "so" target
GNU make 3.82 appears to have a bug which causes problems with our "so"
targets. The "so" targets rely on recursice make calls to build various
subtargets, the shared library and the two executables that link to the shared
library.

On the final recursive make call (to build the executables), GNU make 3.82
wrongly identifies that one dependency (psbase.dev) has changed, and decides it
must rebuild the shared library. By this stage, however, the compiler options
have been setup to build an executable, rather than the shared library.

As a result, we build an executable with the name of the shared lib, and
then try to link that into an executable, which results in linker errors.

As it happens, the use of recursive make calls means we can safely drop a
dependency, which will prevent make from re-evaluating the problematic target,
and thus avoiding the problem arising.

CLUSTER_UNTESTED