Bug 693398: avoid marking/reloc'ing unsuitable pointer
The gx_image_enum object contains a pointer (cie_range) to the CIE range in
the color space, if it's for an image in certain types of CIE space. The
pointer points *into* an area of allocated memory, and not to the "top" of
the allocated block. As a result, when garbage collection tries to relocate,
there isn't a valid header structure, and the values it ends up using for the
relocation are nonsense.

Removing the cie_range pointer from the GC pointer enumeration is not
because the color space structure and it's contents can move. Since this is
a fairly rare case, and the CIE range lookup is trivial, the cie_range
pointer is replaced with a boolean (so we decide *whether* to use the range
only once), and then the range itself is retrieved at the time it's used.

No cluster differences.