pdfwrite - fix bug with 64-bit file offsets to permit large files
Bug #693610 " Regression: pdfwrite seg fault with the 32 bit build
starting with 9b10f8c659ffc643ab71c6180a70da4f39075578"

When doing the work to use 64-bit file offsets, and thus permit files in
excess of 4GB, I missed an instance.When creating the xref table, if an
object is unused (which is rare but possible), we need to scan past it in
the xref offsets temporary file. We do this by calculating the position of
the next object id in the file.

However, instead of using the size of a gp_offset_t (which is what is written
into the file) we were still using the size of an unsigned long. Because
these are the same on a 64-bit build the regular cluster test doesn't catch
this.

Fixed simply by using the correct size. No differences expected, as explained
above this only exhibits on platforms where gp_offset_t and unsigned long
have different sizes.