pdfwrite - alter escapement of text strings in outlines
Bug 693614 "pdfmark: accented character "c" in titles of generated
PDF bookmarks is not displayed properly"

I believe this is actually an Acrobat bug, the text string emits the
Unicode 0x01 0x0d as a binary 0x01 and an escaped character '\n'. This
follows the rulses for text strings. However Acrobat files to display
the Title correctly.

Running the file through Distiller I find that the string is emitted as
0x01 '\015', the octal representation of 0x0d.

This commit alters the escapement of characters in outlines so that they
always end up as octal escapes. This seems to pacify Acrobat.

No differences expected