Bug 693621: Clist logic didn't properly reset after erasepage.
The "normal" page rendering logic works because it flushes the buffer
memory to the clist, then after the page is printed, it clears out
the clist data for the next page. By changing the clist_fillpage to
do this (without actually rendering anything) we now clear out the
clist for a new page. This has the advantage that any clist operations
accumulated prior to the fillpage will be discarded, INCLUDING any
previous fillpage operations which gets rid of duplicate fillpage
actions which can be time consuming, particularly on embedded systems
or printers.