pdfwrite - fix type 1 to type 2 font converter
Bug #693633 "Font character missing in PDF output - appears in other devices and older versions"

Type 2 fonts are required to start the path operations with a moveto class
operation, not a lineto, type 1 fonts do not have this restriction. So when
converting from type 1 to type 2 we must detect if the initial operation is
a 'line' or a 'move' and emit an explcit 'move' in the case where the first
operation is a line.

The code to do this was assuming that if there were operands on the stack
this meant that the glyph width (from the hsbw operator) was present.
However, this is not true if the font contains any hinting operations as
the width must be emitted as part of those hints.

This caused the code to emit confused and invalid type 2 CharStrings.

Fixed by adding a boolean to track whether the stack contains the glyph
width or not, and test that instead of the number of operands.

No differences expected