pdfwrite/ps2write - first part of reworking colour management
This commit provides a new path through the pdfwrite code and uses the ICC
CMS for linework/text, which it did not previously use. The image code already#
used the CMS, at least in part.

There is a 'place holder' switch called PDFUseOldCMS, defaulting to true
at the moment, which controls whether we use the old path or the new.

The new colour code already shows marked improvements with the LeaveColorUnchanged,
in particular /Separation and /Indexed /Device* colour spaces are now preserved
when going to PostScript, which they were not before.

There are also some changes which affect the embedding of ICCBased spaces, and seem
to improve a small number of these.

When enabled this does cause a *large* number of differences, however these
are all either progressions, or very minor differences, invisible to the naked eye.