ps2write - colour conversion work Indexed colour spaces
When dealing with /Indexed colour spaces we concretize to the base space,
if there is no ICC profile associated with the base space then we can't
use it to get the number of colour components to convert, so use the
base space Num_components instead.

Prevents crashes with Indexed spaces with complex base types (for example
/CIEBased spaces)

No differences expected.