ps2write - improve handling of complex colour spaces
Part of the rework on colour handling.

When we write complex spaces (eg /Indexed with a /CIEBased space) and we
use the underlying space (CIEBased) on its own, we only write the underlying
space once, and refer to it form the complex space.

However if we use the underlying space first, then it gets 'resolved', ie
it has all the indirect references replaced. When we then use the complex
space, any references in the underlying space have already been resolved.

Normally this is not a problem because when we try to resolve any indirect
references, we check to see if the object is an indirect reference by testing
a number of its features. This works well but *if* the tint transform is
already resolved, *and* it has fewer than 2 elements, then this would provoke
an error, because the code did not test the length of the array (after already
checking its type) before attempting to extract objects form it.

This patch merely checks the array length and assumes it is not an indirect
reference if the array is too short.

No differences expected.