Bug 693212: Fix strange artifacts with thin lines.
Ghostscript includes special code for 'thin' lines. These are
a fudge at best.

To avoid dropouts in such lines, caused by the end points of various
lines not being drawn, we introduced a further fudge (see bug 687721).
This would restore the end points on lines in some circumstances.
This fudge causes the effect reported in bug 693212.

The correct fix should be to completely remove thin line support from
gs, as at best we are fudging a fudge. Given that this seems unlikely
to happen, we will make do for now by tweaking the fudge introduced
for bug 687721 so that it triggers in fewer cases (specifically, it
will never trigger for horizontal or vertical lines).

This is more in keeping with the intention of that change anyway.

This does produce diffs, but no disastrous ones.