txtwrite - fix fallback to single characters.
There were some serious logical flaws in the Unicode creation, when there
is no Unicode information available and we try to manufacture some.

The glyph name matching was completely broken, and not only failed to match
glyph names but broke the fallback to the original character codes too.

This seems to do a much better job (and no longer crashes either).

No differences expected, we don't test txtwrite