Bug 693115: Fix Smask handling in the interpreter and prevent multiple rendering
The SMask was being rendered before every object, and a subsequent SMask
set with a 'gs' operator (ExtGState) was being rendered with a previous
SMask rather than simply replacing it. By setting the SoftMask to a new
value (neither //null, nor the paramdict) we can still detect that a
SMask is in effect, and thus create transparency groups as needed, but
we no longer render the mask multiple times, improving performance.
In .execmaskgroup, after rendering the SMask into a mask_stack maskbuf,
zap it, using 'where' to find the ExtGState dictionary that holds it.

All differences seen with cluster regression are either progressions
or are very minor except sumatra/VOX.pdf with pdfwrite, but this was
broken previously. Now it's still broken, but differently.