pdfwrite/ps2write - New colour conversion, fix default space detection
When we get a /Indexed space, we check the base space and if it is an ICC
space we further check to see if it is a space which we substituted for
one of the device spaces, and if the current strategy means we should not
convert such a space.

Some missing braces meant that if the test failed we didn't write the colour
space at all.

No differences expected