Bug 693674: introduce memory limit for UFST bitmaps
Previously, UFST had "unlimited" memory for glyph bitmaps. This appears to allow
bitmaps large enough to cause an overflow in the UFST rendering code.

Now there is a limit on the memory UFST can allocate for a bitmap (tied to the
maximum "interim" bitmap Ghostscript will allocate when creating an entry for
the glyph cache).

This has two benefits: prevents the crash by avoiding the rendering overflow in
UFST, and prevents us from pointlessly allocating huge glyph bitmaps in cases
where we really want to render from an outline anyway.

No cluster differences.