Fix indeterminate page background with plank and comparefiles/104-01.ps
This file uses the PS erasepage operator to paint the background to
a non-white value. The mem_planar_fill_rectangle_hl_color used the
'devn' colors without checking the type. With the plank device these
colors were 'pure' or 'ht_binary' so the devn values were bogus.
If the color type was binary, an address was used as color values.
A pure color was OK as long as the number of planes did not exceed
the size of the gx_color_index and the bytes in the color index
corresponded to the planes.

Also add warning (DEBUG build) if mem_planar_strip_tile_rect_devn
ever uses non-devn colors.

Progressions with 104-01.ps, 12-05.PS, 12-07B.PS and 29-07A.PS