Bug 692862: Fix stroking problem with antialiasing
The fix for bug 691228 causes problems in some cases; when joining
two segment that are directly in line with the next, we add a
circular line cap to ensure correct rendering. Which side of this
end cap we end up on depends on which direction the line then
continues in.