pdfwrite - Auto-rotation. Prefer auto-rotate over DSC comments
When auto-rotation is not /None we check both the heuristically determined
auto-rotate value and (for PostScript input) any DSC document or page level
comments.

Previously where these conflicted we used the DSC comments, but this seems
obtuse when the user has asked for auto-rotation. This reverses the default
so that auto-rotation will be preferred.

Also updated the documentation which incorrectly stated that Orientation
DSC comments were ignored.

Expected differences
A few files are now rotated where previously they were not