Bug 693686: Improve heurisic for vertical metrics from FT
Freetype "derives" vertical metrics when a font does not contain any, so we
use a heuristic to decide whether to use them, or not. With the revised FAPI
for use with the non-PS/PDF languages, the "heurstic" was "augmented" with an
unsuitably simple test - so now that simple test has been replaced with
an improved heuristic which works for both PS and PCL.

No cluster differences.