pdfwrite - improve handling of broken fonts
Bug #693711 "pdfwrite segfault"

The specimen file contains two fonts which are badly broken, in one case
the broken data stream decodes to a 'return' operation without a prior
'callsubr'. This caused the SEAC detection code to attempt to free memory
using uninitialised data, which caused a seg fault. We now detect the call
stack depth, and return an error if we return when no subroutine is executing.

This caused a later heap error, as did the other font, because the regular
type 1 font interpreter also didn't like the font data, and returned an error.
This caused us not to write a FontFile resource, but the FontDescriptor
writing code tried to write one anyway. Fixed by checking that there is a
FontFile before trying to write one.

The specimen file now runs to completion, but will not embed the broken
fonts. As a result it will only work 'correctly' on systems which have the
missing fonts available. Although given that these fonts are broken, it
seems unlikely these will work anyway.

No differences expected.