pdfwrite - fix problem with discontinuous xref tables and 64-bit file sizes
Bug #693710 "Regression: Missing output with pdfwrite and 32-bit build"

The update to use 64-bit offsets for file sizes missed a case when finding
the end of an xref section. This only occurs when an object has been reserved
for future use, and then not actually used (eg a Destination for a Link on a
page which is not then produced).

Corrected the type from ulong to gs_offset_t.

No differences expected.