pdfwrite - fix non-standard ranges in CIE spaces converted to ICC wiht linework
Bug #693698 "Wrong color in file with pdfwrite"

The specimen file claims to be Lab but actually isn't. The colour components
are scaled to 0->100, 0->255, 0->255 instead of 0->100, -128 -> 127, -128 -> 127

While the shading and image code scales the input values into 0->1 when emitting
an equivalent ICCBased space, the linework code did not.

This commit addresses that, and also adds code to deal with the same
issue in the new colour code.

There are a few cluster differences but the results are visually identical
or extremely similar.