Bug 693613: add an explicit memory parameter to gs_path_enum_copy_init()
In zpathforall() the path enumerator object is allocated with the current
allocator for the Postscript VM mode. But gs_path_enum_copy_init() used the
memory pointer from the graphics state for all it's allocations, and also to
set the path enumerator's "memory" pointer.

Clearly there was the potential for problems there: in this case, the current
VM mode was global, but the graphics state was allocated in local. In this
scenario, we immediately do a GC on *local* VM after zpathforall() and before
we reach  path_continue(). Because the path enumerator was allocated in global
VM, we skip it during this GC, meaning the copied path and its contents (in
local VM) were not marked, the memory considered free, and overwritten during
the compaction phase.

By adding an explicit memory parameter to gs_path_enum_copy_init() we can
ensure that all the allocations for the path enumerator are consistent, and that
the memory pointer in the enumerator is also correct.

No cluster differences.