pdfwrite - only process pagelabels for pdfwrite device
Bug #693721 "Regression: file cannot be read starting with
5784bfbfba7191cacce5309e88afac0851287460"

The pagelabels are handled as device parameters rather than pdfmarks, because
the pdfmark syntax can't handle the variety of page labels possible in a
PDF file.

Of course only the pdfwrite device can handle this device parameter, so we
need to only perform this if the current device is pdfwrite.

No differences expected, cluster only tests full builds, inclusive of
the pdfwrite device.