Bug 693488: Improved glyph name aliasing.
Implement a more flexible approach to glyph name aliasing which handles all the
names in the Adobe Glyph List.

Thanks to John Wehle ( john@feith.com ) for this implementation.

For the time being, you can use "-dUSEOLDGLYPHALIASING" to revert to the old
code - assuming all goes well, that option will be removed for the release
following 9.09.

No cluster differences.