Fix reading uninitialized point (bug #693655), remove obsolete supported conditional configurations and refactor.
A logical error in paint_path resulted in reading the cursor
uninitialized.  Early in XL development NEWPATH_AFTER_PAINT and
NO_SCALED_DASH_PATTERN were added to support incorrect language in the
specification but their behavior is well established after 3 major
revisions of the PCL-XL language can be safely removed allowing
simplification of this code.

Even with those changes the paint_path procedure was much too complex
to be understandable, it is easy to see how a logical error would go
unnoticed.  We try to refactor the code such that client procedures
take on some of the lifting and get rid of the gotos in hopes of
having something cohesive and correct.