Change the colorSpace header entry in generated profiles to make pdfwrite happy
The generated profiles for certain kinds of colour space were being created
as 3CLR and 4CLR, which is absolutely fine in the general case. However,
the new pdfwrite code uses the internal 'icc_equivalent' profile generated
for a CIEBased PostScript colour space (previously pdfwrite generated its
own profile).

But The PDF Reference says that only Gray, RGB, CMYK and Lab are permissible
values for the header. So here we pretend that the generated profiles are
in either RGB or CMYK, so that the emitted PDF will be valid.

Because the profiles are only used to convert into a CIE device-independent
colour space, which will later be converted into a device space, this 'lie'
should not cause any problems (these spaces are not currently emitted at any
other point, and are never used as device profiles).

No differences expected