pdfwrite - always fallback for type 4 images if we change colour space
Type 4 (chroma keyed) images can't readily have their colour space changed,
if we are going to a space with fewer components. For example, if we convert
a CMYK image to Gray, then many different CMYK values will map to the same
gray value, leading to parts of the image being masked which should not be.

Technically it is possible to change colour space the other way, but it is
easier for now to simply cause all such images to use the fallback code if
we are changing colour space.