pdfwrite - prevent dereference of NULL pointer when no BG/UCR info present
Bug #693728 " Incompatible and large PDF output, and a crash with -dUCRandBGInfo=/Preserve"

The input is XPS and XPS has no concept of BG/UCR, so the pointer to the
information is NULL. If we had set the parameter -dUCRandBGInfo=/Preserve
then we would attempt to dereference the NULL Pointer.

We now check the pointer before attempting to dereference it.

No differences expected.