Memory squeezing fix
If ialloc_init fails part way through, we attempt to free the allocated
objects, but we do so incorrectly as the object we get passed back is
actually internal to the allocation that was made (and there were actually
multiple allocations).

We add a function to do this properly, together with a stern warning
that this should only be used for allocation error cleanup.