Add an ID to memory allocations.
This is to aid following objects through GC.

This is fairly straight forward with gs_struct_.... type allocations where
there is a memory manager header structure associated with the allocation,
so the ID is added to that.

But so called "string" allocations (i_alloc_string() and co) are more tricky
because these have no header, so these have the ID "tacked on" at the beginning
of the memory allocation. This means that the memory manager and the garbage
collector need to "just know" that the extra bytes exist.

The assigning and incrementing of the ID variable (normally) on happen in a
debug build. The ID variable itself can be left out by adding:
-DGS_USE_MEMORY_HEADER_ID=0

(or equivalent) to the compiler command line.

Finally, in a debug build, there is a get_mem_hdr_id() convenience function
you can call from a debugger - it takes a pointer to block of memory, and
returns the ID assigned to the allocation. Obviously, if the function is called
on memory automatically allocaed on the stack, it will probably cause bad
things to happen.

No cluster differences.